eyeguard-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
firebase-admin==6.5.0
google-generativeai==0.7.2